refactor voting_round out to own module
